+++ packages/neuron-ui/src/states/initStates/settings.ts
@@ -1,7 +1,10 @@
 import { addressBook, wallets, networks } from 'services/localCache'
 
 export const settingsState: State.Settings = {
-  showAddressBook: addressBook.isVisible(),
+  general: {
+    skipDataAndType: false,
+    showAddressBook: addressBook.isVisible(),
+  },
   networks: networks.load(),
   wallets: wallets.load(),
 }