+++ packages/neuron-ui/src/states/stateProvider/actionCreators/app.ts
@@ -26,6 +26,7 @@ export const initAppState = () => (dispatch: StateDispatch, history: any) => {
           syncedBlockNumber = '',
           connectionStatus = false,
           codeHash = '',
+          skipDataAndType = false,
         } = res.result
         dispatch({
           type: NeuronWalletActions.InitAppState,
@@ -38,6 +39,7 @@ export const initAppState = () => (dispatch: StateDispatch, history: any) => {
             syncedBlockNumber,
             connectionStatus,
             codeHash,
+            skipDataAndType,
           },
         })
         if (!wallet) {