+++ packages/neuron-ui/src/types/Controller/index.d.ts
@@ -70,4 +70,5 @@ declare namespace Controller {
     hash: string
     description: string
   }
+  type SetSkipAndTypeParam = boolean
 }