+++ packages/neuron-wallet/src/controllers/index.ts
@@ -3,6 +3,7 @@ import NetworksController from './networks'
 import WalletsController from './wallets'
 import TransactionsController from './transactions'
 import SyncInfoController from './sync-info'
+import SkipDataAndTypeController from './skip-data-and-type'
 
 export default {
   AppController,
@@ -10,4 +11,5 @@ export default {
   WalletsController,
   TransactionsController,
   SyncInfoController,
+  SkipDataAndTypeController,
 }