+++ packages/neuron-wallet/src/database/chain/ormconfig.ts
@@ -11,6 +11,7 @@ import Output from './entities/output'
 import SyncInfo from './entities/sync-info'
 import { InitMigration1566959757554 } from './migrations/1566959757554-InitMigration'
 import { AddTypeAndHasData1567144517514 } from './migrations/1567144517514-AddTypeAndHasData'
+import { ChangeHasDataDefault1568621556467 } from './migrations/1568621556467-ChangeHasDataDefault'
 
 export const CONNECTION_NOT_FOUND_NAME = 'ConnectionNotFoundError'
 
@@ -31,7 +32,7 @@ const connectOptions = async (genesisBlockHash: string): Promise<SqliteConnectio
     type: 'sqlite',
     database,
     entities: [Transaction, Input, Output, SyncInfo],
-    migrations: [InitMigration1566959757554, AddTypeAndHasData1567144517514],
+    migrations: [InitMigration1566959757554, AddTypeAndHasData1567144517514, ChangeHasDataDefault1568621556467],
     logging,
   }
 }