+++ packages/neuron-wallet/src/startup/sync-block-task/init-database.ts
@@ -2,6 +2,7 @@ import initConnection from 'database/chain/ormconfig'
 import Utils from 'services/sync/utils'
 import { updateMetaInfo, getMetaInfo } from 'database/chain/meta-info'
 import LockUtils from 'models/lock-utils'
+import logger from 'utils/logger'
 import genesisBlockHash from './genesis'
 
 export const initDatabase = async () => {
@@ -11,6 +12,7 @@ export const initDatabase = async () => {
     const systemScriptInfo = await LockUtils.systemScript()
     updateMetaInfo({ genesisBlockHash: hash, systemScriptInfo })
   } catch (err) {
+    logger.debug('initDatabase error:', err)
     try {
       const metaInfo = getMetaInfo()
       await initConnection(metaInfo.genesisBlockHash)